platform_stage/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Deprecated==1.2.13
dj-database-url==2.1.0
Django==5.0.1
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-google-maps==0.13.0
django-js-asset==2.1.0
django-redis==5.4.0
django-taggit==5.0.1
django-user-visit==2.0
gunicorn==21.2.0
importlib-metadata==7.0.0
mysqlclient==2.2.1
packaging==23.2
psycopg2==2.9.9
pyparsing==3.1.1
pytz==2023.3.post1
redis==5.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
ua-parser==0.18.0
user-agents==2.2.0
whitenoise==6.6.0
wrapt==1.16.0
zipp==3.17.0
